+++ packages/cli/.gitignore
@@ -1,4 +1,4 @@
 bin
 codegen
 out.txt
-artifacts
+out
