+++ packages/cli/test/artifacts/polkadot-metadata.scale
