+++ packages/cli/test/artifacts/test_descriptors-altered.json
@@ -1,7 +1,7 @@
 {
   "polkadot": {
-    "metadata": "polkadot-metadata.scale",
-    "outputFolder": "artifacts/descriptors",
+    "metadata": "test/artifacts/polkadot-metadata-altered.scale",
+    "outputFolder": "out",
     "descriptors": {
       "System": {
         "constants": {
