+++ packages/cli/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.base",
-  "include": ["src", "tests", "artifacts"],
+  "include": ["src", "tests", "out"],
   "compilerOptions": {
     "baseUrl": "src",
     "noUnusedLocals": false,
